Use int32 rather than uint32 as behaviour on Java is tricky
https://developers.google.com/protocol-buffers/docs/proto#scalar says "In
Java, unsigned 32-bit and 64-bit integers are represented using their
signed counterparts, with the top bit simply being stored in the sign bit."

...which means that it is necessary (and tricky) to check the bounds of
unsigned values when deserialising a Protocol Buffer under Java.

We don't need the extra precision so let's not require it. Moreover, port
numbers are unsigned shorts but the smallest integer type in Protocol
Buffers is 32bits wide.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>